requirements.txt
Pillow 
PyYAML